// src/app/demo/models/modelsgeneral/cargoviewmodel.ts
export class Cargo{
    codigo?: number;
    carg_Id?: number;
    carg_Descripcion?:string;
    usua_Creacion?:number;
    usua_Modificacion?: number;
    usuaCreacion?:string;
    usuaModificacion?:string;
    carg_FechaCreacion?: string;
    carg_FechaModificacion?: string;
    carg_Estado?: boolean;

}